lambda/packages/fraud_alert_lambda_package/src/config.py
import os
from dataclasses import dataclass
from dotenv import load_dotenv

load_dotenv()


@dataclass(frozen=True)
class Settings:
    transactions_table: str = os.getenv("TRANSACTIONS_TABLE", "transactions")
    accounts_table: str = os.getenv("ACCOUNTS_TABLE", "accounts")
    twilio_account_sid: str = os.getenv("TWILIO_ACCOUNT_SID", "")
    twilio_auth_token: str = os.getenv("TWILIO_AUTH_TOKEN", "")
    twilio_whatsapp_from: str = os.getenv("TWILIO_WHATSAPP_FROM", "whatsapp:+14155238886")


settings = Settings()