lambda/packages/fraud_alert_lambda_package/requirements.txt
twilio==9.10.3
python-dotenv==1.2.2